+++ ante/cosmos/eip712.go
@@ -1,8 +1,8 @@
 package cosmos
 
 import (
+	"context"
 	"fmt"
-	"strconv"
 
 	ethcrypto "github.com/ethereum/go-ethereum/crypto"
 	"github.com/ethereum/go-ethereum/crypto/secp256k1"
@@ -11,8 +11,12 @@ import (
 	anteinterfaces "github.com/cosmos/evm/ante/interfaces"
 	"github.com/cosmos/evm/crypto/ethsecp256k1"
 	"github.com/cosmos/evm/ethereum/eip712"
+	evmtypes "github.com/cosmos/evm/x/vm/types"
 
+	txv1beta1 "cosmossdk.io/api/cosmos/tx/v1beta1"
 	errorsmod "cosmossdk.io/errors"
+	txsigning "cosmossdk.io/x/tx/signing"
+	"cosmossdk.io/x/tx/signing/aminojson"
 
 	"github.com/cosmos/cosmos-sdk/codec"
 	codectypes "github.com/cosmos/cosmos-sdk/codec/types"
@@ -21,7 +25,6 @@ import (
 	errortypes "github.com/cosmos/cosmos-sdk/types/errors"
 	"github.com/cosmos/cosmos-sdk/types/tx/signing"
 	authante "github.com/cosmos/cosmos-sdk/x/auth/ante"
-	"github.com/cosmos/cosmos-sdk/x/auth/migrations/legacytx"
 	authsigning "github.com/cosmos/cosmos-sdk/x/auth/signing"
 )
 
@@ -136,6 +139,7 @@ func (svd LegacyEip712SigVerificationDecorator) AnteHandle(ctx sdk.Context,
 		ChainID:       chainID,
 		AccountNumber: accNum,
 		Sequence:      acc.GetSequence(),
+		Address:       acc.GetAddress().String(),
 	}
 
 	if simulate {
@@ -177,22 +181,41 @@ func VerifySignature(
 			return errorsmod.Wrap(errortypes.ErrNoSignatures, "tx doesn't contain any msgs to verify signature")
 		}
 
-		txBytes := legacytx.StdSignBytes( //nolint:staticcheck // TODO: fix
-			signerData.ChainID,
-			signerData.AccountNumber,
-			signerData.Sequence,
-			tx.GetTimeoutHeight(),
-			legacytx.StdFee{
-				Amount: tx.GetFee(),
-				Gas:    tx.GetGas(),
-			},
-			msgs, tx.GetMemo(),
-		)
-
-		signerChainID, err := strconv.ParseUint(signerData.ChainID, 10, 64)
+		anyMsgs, err := eip712.ToAnyMsgs(msgs)
 		if err != nil {
-			return errorsmod.Wrapf(err, "failed to parse chain-id: %s", signerData.ChainID)
+			return err
 		}
+		feeAmount := eip712.ToFeeAmount(tx.GetFee())
+		txData := txsigning.TxData{
+			Body: &txv1beta1.TxBody{
+				Messages:      anyMsgs,
+				Memo:          tx.GetMemo(),
+				TimeoutHeight: tx.GetTimeoutHeight(),
+			},
+			AuthInfo: &txv1beta1.AuthInfo{
+				Fee: &txv1beta1.Fee{
+					Amount:   feeAmount,
+					GasLimit: tx.GetGas(),
+				},
+			},
+		}
+		signModeHandler := aminojson.NewSignModeHandler(aminojson.SignModeHandlerOptions{})
+		signer := txsigning.SignerData{
+			ChainID:       signerData.ChainID,
+			AccountNumber: signerData.AccountNumber,
+			Sequence:      signerData.Sequence,
+			Address:       signerData.Address,
+		}
+		txBytes, err := signModeHandler.GetSignBytes(context.Background(), signer, txData)
+		if err != nil {
+			return errorsmod.Wrap(err, "failed to get sign bytes using aminojson")
+		}
+
+		ethChainID := evmtypes.GetEthChainConfig().ChainID
+		if ethChainID == nil {
+			return errorsmod.Wrap(errortypes.ErrInvalidChainID, "eth chain ID not configured")
+		}
+		signerChainID := ethChainID.Uint64()
 
 		txWithExtensions, ok := tx.(authante.HasExtensionOptionsTx)
 		if !ok {