+++ ethereum/eip712/encoding_legacy.go
@@ -1,12 +1,19 @@
 package eip712
 
 import (
+	"context"
 	"encoding/json"
 	"errors"
 	"fmt"
 
 	apitypes "github.com/ethereum/go-ethereum/signer/core/apitypes"
 
+	txv1beta1 "cosmossdk.io/api/cosmos/tx/v1beta1"
+	errorsmod "cosmossdk.io/errors"
+	txsigning "cosmossdk.io/x/tx/signing"
+	"cosmossdk.io/x/tx/signing/aminojson"
+
+	cryptotypes "github.com/cosmos/cosmos-sdk/crypto/types"
 	sdk "github.com/cosmos/cosmos-sdk/types"
 	txTypes "github.com/cosmos/cosmos-sdk/types/tx"
 	"github.com/cosmos/cosmos-sdk/x/auth/migrations/legacytx"
@@ -157,14 +164,12 @@ func legacyDecodeProtobufSignDoc(signDocBytes []byte, eip155ChainID uint64) (api
 
 	// Use first message for fee payer and type inference
 	msg := msgs[0]
-
 	signerInfo := authInfo.SignerInfos[0]
-
-	stdFee := &legacytx.StdFee{
-		Amount: authInfo.Fee.Amount,
-		Gas:    authInfo.Fee.GasLimit,
+	var pubKey cryptotypes.PubKey
+	err := protoCodec.UnpackAny(signerInfo.PublicKey, &pubKey)
+	if err != nil {
+		return apitypes.TypedData{}, errorsmod.Wrap(err, "failed to unpack signer public key")
 	}
-
 	signers, _, err := protoCodec.GetMsgV1Signers(msg)
 	if err != nil {
 		return apitypes.TypedData{}, err
@@ -175,15 +180,36 @@ func legacyDecodeProtobufSignDoc(signDocBytes []byte, eip155ChainID uint64) (api
 	}
 
 	// WrapTxToTypedData expects the payload as an Amino Sign Doc
-	signBytes := legacytx.StdSignBytes( //nolint:staticcheck // TODO: fix
-		signDoc.ChainId,
-		signDoc.AccountNumber,
-		signerInfo.Sequence,
-		body.TimeoutHeight,
-		*stdFee,
-		msgs,
-		body.Memo,
-	)
+	anyMsgs, err := ToAnyMsgs(msgs)
+	if err != nil {
+		return apitypes.TypedData{}, err
+	}
+	feeAmount := ToFeeAmount(authInfo.Fee.Amount)
+	txData := txsigning.TxData{
+		Body: &txv1beta1.TxBody{
+			Messages:      anyMsgs,
+			Memo:          body.Memo,
+			TimeoutHeight: body.TimeoutHeight,
+		},
+		AuthInfo: &txv1beta1.AuthInfo{
+			Fee: &txv1beta1.Fee{
+				Amount:   feeAmount,
+				GasLimit: authInfo.Fee.GasLimit,
+			},
+		},
+	}
+
+	signModeHandler := aminojson.NewSignModeHandler(aminojson.SignModeHandlerOptions{})
+	signer := txsigning.SignerData{
+		ChainID:       signDoc.ChainId,
+		AccountNumber: signDoc.AccountNumber,
+		Sequence:      signerInfo.Sequence,
+		Address:       pubKey.Address().String(),
+	}
+	signBytes, err := signModeHandler.GetSignBytes(context.Background(), signer, txData)
+	if err != nil {
+		return apitypes.TypedData{}, errorsmod.Wrap(err, "failed to get sign bytes using aminojson")
+	}
 
 	typedData, err := LegacyWrapTxToTypedData(
 		protoCodec,